Move room upgrading to the roomserver
